requirements.txt
flask 
Flask-Bcrypt
Flask-Cors
Flask-Mail
poetry
pymongo 
python-dotenv
requests
midtransclient 
gunicorn
pyJWT